created accounts successfully

+++ Banking-app/src/main/java/com/bankingApp/Banking/app/Repository/AccountRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.bankingApp.Banking.app.Entity.Accounts;
 
-public interface AccountRepo extends JpaRepository<Accounts,Integer> {
+public interface AccountRepo extends JpaRepository<Accounts,Long> {
     
 }
